+++ src/mantarray_desktop_app/server.py
@@ -57,8 +57,9 @@ from flask import Response
 from flask_cors import CORS
 from flask_socketio import SocketIO
 from immutabledict import immutabledict
-from mantarray_file_manager import MANTARRAY_NICKNAME_UUID
-from mantarray_waveform_analysis import CENTIMILLISECONDS_PER_SECOND
+from pulse3D.constants import CENTIMILLISECONDS_PER_SECOND
+from pulse3D.constants import MANTARRAY_NICKNAME_UUID
+from pulse3D.constants import METADATA_UUID_DESCRIPTIONS
 import requests
 from semver import VersionInfo
 from stdlib_utils import drain_queue
@@ -72,7 +73,6 @@ from .constants import DEFAULT_SERVER_PORT_NUMBER
 from .constants import GENERIC_24_WELL_DEFINITION
 from .constants import INSTRUMENT_INITIALIZING_STATE
 from .constants import LIVE_VIEW_ACTIVE_STATE
-from .constants import METADATA_UUID_DESCRIPTIONS
 from .constants import MICRO_TO_BASE_CONVERSION
 from .constants import MICROSECONDS_PER_CENTIMILLISECOND
 from .constants import MICROSECONDS_PER_MILLISECOND