+++ tests/data_analyzer/test_beta_1_mode.py
@@ -7,6 +7,7 @@ import time
 
 from freezegun import freeze_time
 from mantarray_desktop_app import CONSTRUCT_SENSOR_SAMPLING_PERIOD
+from mantarray_desktop_app import data_analyzer
 from mantarray_desktop_app import DATA_ANALYZER_BUFFER_SIZE_CENTIMILLISECONDS
 from mantarray_desktop_app import FIFO_READ_PRODUCER_DATA_OFFSET
 from mantarray_desktop_app import FIFO_READ_PRODUCER_SAWTOOTH_PERIOD
@@ -20,11 +21,11 @@ from mantarray_desktop_app import REFERENCE_SENSOR_SAMPLING_PERIOD
 from mantarray_desktop_app import ROUND_ROBIN_PERIOD
 from mantarray_desktop_app import STOP_MANAGED_ACQUISITION_COMMUNICATION
 from mantarray_desktop_app import TIMESTEP_CONVERSION_FACTOR
-from mantarray_waveform_analysis import BUTTERWORTH_LOWPASS_30_UUID
-from mantarray_waveform_analysis import CENTIMILLISECONDS_PER_SECOND
-from mantarray_waveform_analysis import Pipeline
-from mantarray_waveform_analysis import PipelineTemplate
+from mantarray_desktop_app.data_analyzer import get_force_signal
 import numpy as np
+from pulse3D.constants import BUTTERWORTH_LOWPASS_30_UUID
+from pulse3D.constants import CENTIMILLISECONDS_PER_SECOND
+from pulse3D.transforms import create_filter
 import pytest
 from scipy import signal
 from stdlib_utils import drain_queue
@@ -72,14 +73,11 @@ def fill_da_input_data_queue(input_queue, num_seconds):
             input_queue.put_nowait(ref_packet)
 
 
-# TODO Tanner (7/15/21): Should eventually add the following 3 tests for Beta 2 mode
-
-
 @pytest.mark.slow
 def test_DataAnalyzerProcess_beta_1_performance__fill_data_analysis_buffer(
     runnable_four_board_analyzer_process,
 ):
-    # 8 seconds of data (625 Hz) coming in from File Writer to going back to Main
+    # 11 seconds of data (625 Hz) coming in from File Writer to going through to Main
     #
     # mantarray-waveform-analysis v0.3:     4.148136512
     # mantarray-waveform-analysis v0.3.1:   3.829136133
@@ -89,6 +87,8 @@ def test_DataAnalyzerProcess_beta_1_performance__fill_data_analysis_buffer(
     # 30 Hz Butterworth filter:             2.935009033  # Tanner (9/10/20): not intended to speed anything up, just adding this to show it had it didn't have much affect on performance
     #
     # added twitch metric analysis:         3.013469479
+    # initial pulse3D import:               3.855403546
+    # pulse3D 0.23.3:                       3.890723909
 
     p, board_queues, comm_from_main_queue, comm_to_main_queue, _ = runnable_four_board_analyzer_process
     put_object_into_queue_and_raise_error_if_eventually_still_empty(
@@ -107,7 +107,7 @@ def test_DataAnalyzerProcess_beta_1_performance__fill_data_analysis_buffer(
     drain_queue(board_queues[0][1])
     drain_queue(comm_to_main_queue)
 
-    # print(f"Duration (seconds): {dur_seconds}")  # pylint:disable=wrong-spelling-in-comment # Eli (4/8/21): this is commented code that is deliberately kept in the codebase since it is often toggled on/off during optimization
+    # print(f"Duration (seconds): {dur_seconds}")  # Eli (4/8/21): this is commented code that is deliberately kept in the codebase since it is often toggled on/off during optimization
     assert dur_seconds < 10
 
 
@@ -115,10 +115,12 @@ def test_DataAnalyzerProcess_beta_1_performance__fill_data_analysis_buffer(
 def test_DataAnalyzerProcess_beta_1_performance__first_second_of_data_with_analysis(
     runnable_four_board_analyzer_process,
 ):
-    # Fill data analysis buffer with 7 seconds of data to start metric analysis,
+    # Fill data analysis buffer with 10 seconds of data to start metric analysis,
     # Then record duration of sending 1 additional second of data
     #
-    # start:                                 0.547285524
+    # start:                                0.547285524
+    # initial pulse3D import:               0.535316489
+    # pulse3D 0.23.3:                       0.535428579
 
     p, board_queues, comm_from_main_queue, comm_to_main_queue, _ = runnable_four_board_analyzer_process
     put_object_into_queue_and_raise_error_if_eventually_still_empty(
@@ -141,18 +143,20 @@ def test_DataAnalyzerProcess_beta_1_performance__first_second_of_data_with_analy
     drain_queue(board_queues[0][1])
     drain_queue(comm_to_main_queue)
 
-    # print(f"Duration (seconds): {dur_seconds}")  # pylint:disable=wrong-spelling-in-comment # Eli (4/8/21): this is commented code that is deliberately kept in the codebase since it is often toggled on/off during optimization
+    # print(f"Duration (seconds): {dur_seconds}")  # Eli (4/8/21): this is commented code that is deliberately kept in the codebase since it is often toggled on/off during optimization
     assert dur_seconds < 2
 
 
 @pytest.mark.slow
-def test_DataAnalyzerProcess_beta_1_performance__single_data_packet_per_well(
+def test_DataAnalyzerProcess_beta_1_performance__single_data_packet_per_well_without_analysis(
     runnable_four_board_analyzer_process,
 ):
-    # 1 second of data (625 Hz) coming in from File Writer to going back to Main
+    # 1 second of data (625 Hz) coming in from File Writer to going through to Main
     #
-    # start:                                 0.530731389
-    # added twitch metric analysis:          0.578328276
+    # start:                                0.530731389
+    # added twitch metric analysis:         0.578328276
+    # initial pulse3D import:               0.533860423
+    # pulse3D 0.23.3:                       0.539447351
 
     p, board_queues, comm_from_main_queue, comm_to_main_queue, _ = runnable_four_board_analyzer_process
     put_object_into_queue_and_raise_error_if_eventually_still_empty(
@@ -171,7 +175,7 @@ def test_DataAnalyzerProcess_beta_1_performance__single_data_packet_per_well(
     drain_queue(board_queues[0][1])
     drain_queue(comm_to_main_queue)
 
-    # print(f"Duration (seconds): {dur_seconds}")  # pylint:disable=wrong-spelling-in-comment # Eli (4/8/21): this is commented code that is deliberately kept in the codebase since it is often toggled on/off during optimization
+    # print(f"Duration (seconds): {dur_seconds}")  # Eli (4/8/21): this is commented code that is deliberately kept in the codebase since it is often toggled on/off during optimization
     assert dur_seconds < 2
 
 
@@ -419,9 +423,9 @@ def test_DataAnalyzerProcess__dumps_all_data_when_buffer_is_full_and_clears_buff
 
     # Tanner (6/16/20): The tiny amount of data used in this test doesn't work with mantarray_waveform_analysis functions, so we can mock them to prevent errors
     mocked_force_vals = [np.array([expected_x_vals, expected_y_vals[i]]) for i in range(24)]
-    mocked_compressed_force = mocker.patch.object(
-        Pipeline,
-        "get_compressed_force",
+    mocked_get_force = mocker.patch.object(
+        data_analyzer,
+        "get_force_signal",
         autospec=True,
         side_effect=mocked_force_vals,
     )
@@ -468,7 +472,7 @@ def test_DataAnalyzerProcess__dumps_all_data_when_buffer_is_full_and_clears_buff
     assert data_buffer[23]["construct_data"] is None
     assert data_buffer[23]["ref_data"] is None
 
-    assert mocked_compressed_force.call_count == 24
+    assert mocked_get_force.call_count == 24
 
 
 @freeze_time("2020-06-1 13:45:30.123456")
@@ -533,15 +537,11 @@ def test_DataAnalyzerProcess__create_outgoing_data__normalizes_and_flips_raw_dat
         [test_data[0], (test_data[1] - max(test_data[1])) * -1],
         dtype=np.int32,
     )
-
-    pt = PipelineTemplate(
-        is_beta_1_data=True,
-        noise_filter_uuid=BUTTERWORTH_LOWPASS_30_UUID,
-        tissue_sampling_period=ROUND_ROBIN_PERIOD * MICROSECONDS_PER_CENTIMILLISECOND,
+    filter_coefficients = create_filter(
+        BUTTERWORTH_LOWPASS_30_UUID,
+        ROUND_ROBIN_PERIOD * MICROSECONDS_PER_CENTIMILLISECOND,
     )
-    pipeline = pt.create_pipeline()
-    pipeline.load_raw_gmr_data(normalized_data, np.zeros(normalized_data.shape))
-    expected_compressed_data = pipeline.get_compressed_force()
+    expected_compressed_data = get_force_signal(normalized_data, filter_coefficients, is_beta_2_data=False)
     np.testing.assert_equal(actual[0]["x_data_points"], expected_compressed_data[0, :])
     np.testing.assert_equal(
         actual[0]["y_data_points"], expected_compressed_data[1, :] * MICRO_TO_BASE_CONVERSION